Turned the form skeleton into it's own component. Still needs more to be truly dynamic
